Create README.md
# OIBSIP-  Oasis Infobyte Intern Program
OASIS INFOBYTE offers an online internship of variety of Graphic design. The internship program is designed to provide practical knowledge and experience in the field of Graphic design. Got the chance for 1 month internship of Graphic design. This internship is a great opportunity for student and professionals to enhance their skills and gain hands-on experience. The program is flexible and can be completed from anywhere with an internet connection. Glad to be a part of OASIS INFOBYTE internship program